Refactor: 절대경로, ts file 변경
