+++ backend/src/api/routes/career/index.ts
@@ -1,6 +1,5 @@
 import express from 'express';
-
-import careerController from '../../controller/career/careerController.js';
+import careerController from '@src/controller/career/careerController';
 
 const careerRouter = express.Router();
