+++ backend/src/loaders/connect/database.ts
@@ -1,9 +1,8 @@
 import mongoose from 'mongoose'; // mongodb 연결을 위한 패키지지
+import logger from '@src/config/logger'; // winston 로그 설정
 
-import logger from '../config/winston.js'; // winston 로그 설정
-
-export const dbConnect = (info) => {
-  return new Promise((resolve, reject) => {
+export const dbConnect = (info: any) => {
+  return new Promise<void>((resolve, reject) => {
     mongoose
       .connect(`mongodb://${info}?authSource=admin&authMechanism=SCRAM-SHA-1`, {
         useNewUrlParser: true,
